Вынес тестируемые функции в utils. Не работает подключение utils к файлу с тестами
