+++ js/data/game-data.test.js
@@ -1,40 +1,5 @@
 import assert from 'assert';
-// Откуда берётся assert? (ну т.е. она с mocha устанавливается? Не вижу её в lib)
-// Где посмотреть набор методов в assert. Equal, deepEqual и тд
-// Как подключить сюда файл, в котором находится тестируемая функия
-
-// варианты ответов
-// const WRONG_ANSWER = 0; //пока не используется
-const RIGHT_ANSWER = 1;
-const FAST_RIGHT_ANSWER = 3;
-const SLOW_RIGHT_ANSWER = 4;
-const COUNT_QUESTIONS = 10;
-
-const calculateScore = (answers, healthCount) => {
-
-  if ((answers.length !== COUNT_QUESTIONS) || (healthCount <= 0)) {
-    return -1;
-  }
-
-  let resultScore = answers.reduce((score, answer) => {
-    switch (answer) {
-      case RIGHT_ANSWER:
-        score += 100;
-        break;
-      case FAST_RIGHT_ANSWER:
-        score += 150;
-        break;
-      case SLOW_RIGHT_ANSWER:
-        score += 50;
-        break;
-    }
-    return score;
-  }, 0);
-
-  resultScore += healthCount * 50;
-
-  return resultScore;
-};
+import {calculateScore, Timer} from '../utils';
 
 describe(`Тестирование функции подсчета очков`, () => {
   it(`Если игрок ответил меньше чем на 10 ответов, то очки не начисляются`, () => {
@@ -60,23 +25,7 @@ describe(`Тестирование функции подсчета очков`, 
 });
 
 
-const Timer = function (time) {
-  if (time > 0 && time < Number.MAX_SAFE_INTEGER) {
-    this.time = time;
-  } else {
-    this.time = 0;
-  }
-};
 
-Timer.prototype.tick = function () {
-  this.time -= 1;
-  if (this.time <= 0) {
-    this.time = 0;
-    return 0;
-  } else {
-    return this.time;
-  }
-};
 
 
 describe(`Тестирование функции создания таймера`, () => {
