+++ package.json
@@ -7,8 +7,8 @@
     "eslint": "eslint js/",
     "build": "gulp build",
     "start": "gulp serve",
-    "unitests": "gulp test",
-    "test": "npm run eslint && npm run unitTests"
+    "unit-tests": "gulp test",
+    "test": "npm run eslint && npm run unit-tests"
   },
   "devDependencies": {
     "autoprefixer": "6.7.7",
